line: dump the entire buffer, avoid copies when serializing
Put a small header at the beginning of the 'file' which
points to the metadata. After that dump the raw buffer.
When reading, we read the header and then 'shift' the
buffer to get back the original buffer containing the
line. Since it is not possible to obtain an XBuffer from
an existing buffer without copying, we have to employ a
hack for now.

Signed-off-by: Szymon Łopaciuk <szymon@lopaciuk.eu>